Rename all the strategies to use Default prefix instead of Impl suffix
