+++ core/retries/src/main/java/software/amazon/awssdk/retries/LegacyRetryStrategy.java
@@ -20,7 +20,7 @@ import software.amazon.awssdk.annotations.SdkPublicApi;
 import software.amazon.awssdk.annotations.ThreadSafe;
 import software.amazon.awssdk.retries.api.BackoffStrategy;
 import software.amazon.awssdk.retries.api.RetryStrategy;
-import software.amazon.awssdk.retries.internal.LegacyRetryStrategyImpl;
+import software.amazon.awssdk.retries.internal.DefaultLegacyRetryStrategy;
 import software.amazon.awssdk.retries.internal.circuitbreaker.TokenBucketStore;
 
 /**
@@ -59,7 +59,7 @@ public interface LegacyRetryStrategy extends RetryStrategy<LegacyRetryStrategy.B
      * </pre>
      */
     static Builder builder() {
-        return LegacyRetryStrategyImpl
+        return DefaultLegacyRetryStrategy
             .builder()
             .maxAttempts(DefaultRetryStrategy.Legacy.MAX_ATTEMPTS)
             .tokenBucketStore(TokenBucketStore
