+++ core/retries/src/main/java/software/amazon/awssdk/retries/internal/DefaultStandardRetryStrategy.java
@@ -21,7 +21,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.function.Predicate;
 import software.amazon.awssdk.annotations.SdkInternalApi;
-import software.amazon.awssdk.retries.LegacyRetryStrategy;
+import software.amazon.awssdk.annotations.ToBuilderIgnoreField;
+import software.amazon.awssdk.retries.StandardRetryStrategy;
 import software.amazon.awssdk.retries.api.AcquireInitialTokenRequest;
 import software.amazon.awssdk.retries.api.AcquireInitialTokenResponse;
 import software.amazon.awssdk.retries.api.BackoffStrategy;
@@ -41,31 +42,25 @@ import software.amazon.awssdk.utils.Logger;
 import software.amazon.awssdk.utils.Validate;
 
 /**
- * Implementation of the {@link LegacyRetryStrategy} interface.
+ * Implementation of the {@link StandardRetryStrategy} interface.
  */
 @SdkInternalApi
-public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
-    private static final Logger LOG = Logger.loggerFor(LegacyRetryStrategyImpl.class);
+public final class DefaultStandardRetryStrategy implements StandardRetryStrategy {
+    private static final Logger LOG = Logger.loggerFor(DefaultStandardRetryStrategy.class);
 
     private final List<Predicate<Throwable>> predicates;
     private final int maxAttempts;
     private final boolean circuitBreakerEnabled;
     private final BackoffStrategy backoffStrategy;
-    private final BackoffStrategy throttlingBackoffStrategy;
     private final int exceptionCost;
-    private final int throttlingExceptionCost;
-    private final Predicate<Throwable> treatAsThrottling;
     private final TokenBucketStore tokenBucketStore;
 
-    private LegacyRetryStrategyImpl(Builder builder) {
+    private DefaultStandardRetryStrategy(Builder builder) {
         this.predicates = Collections.unmodifiableList(Validate.paramNotNull(builder.predicates, "predicates"));
-        this.maxAttempts = Validate.isPositive(Validate.paramNotNull(builder.maxAttempts, "maxAttempts"), "maxAttempts");
-        this.circuitBreakerEnabled = builder.circuitBreakerEnabled == null || builder.circuitBreakerEnabled;
+        this.maxAttempts = Validate.isPositive(builder.maxAttempts, "maxAttempts");
+        this.circuitBreakerEnabled = builder.circuitBreakerEnabled;
         this.backoffStrategy = Validate.paramNotNull(builder.backoffStrategy, "backoffStrategy");
-        this.throttlingBackoffStrategy = Validate.paramNotNull(builder.throttlingBackoffStrategy, "throttlingBackoffStrategy");
-        this.exceptionCost = Validate.paramNotNull(builder.exceptionCost, "exceptionCost");
-        this.throttlingExceptionCost = Validate.paramNotNull(builder.throttlingExceptionCost, "throttlingExceptionCost");
-        this.treatAsThrottling = Validate.paramNotNull(builder.treatAsThrottling, "treatAsThrottling");
+        this.exceptionCost = builder.exceptionCost;
         this.tokenBucketStore = Validate.paramNotNull(builder.tokenBucketStore, "tokenBucketStore");
     }
 
@@ -79,7 +74,7 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
     @Override
     public RefreshRetryTokenResponse refreshRetryToken(RefreshRetryTokenRequest request) {
         DefaultRetryToken token = asStandardRetryToken(request.token());
-        AcquireResponse acquireResponse = requestAcquireCapacity(request, token);
+        AcquireResponse acquireResponse = requestAcquireCapacity(token);
 
         // Check if we meet the preconditions needed for retrying. These will throw if the expected condition is not meet.
         // 1) is retryable?
@@ -91,7 +86,12 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
 
         // Refresh the retry token and compute the backoff delay.
         DefaultRetryToken refreshedToken = refreshToken(request, acquireResponse);
-        Duration finalDelay = computeBackoff(request, refreshedToken);
+        Duration backoff = backoffStrategy.computeDelay(refreshedToken.attempt());
+
+        // Take the max delay between the suggested delay and the backoff delay.
+        Duration suggested = request.suggestedDelay().orElse(Duration.ZERO);
+        Duration finalDelay = maxOf(suggested, backoff);
+
         logRefreshTokenSuccess(refreshedToken, acquireResponse, finalDelay);
         return RefreshRetryTokenResponseImpl.create(refreshedToken, finalDelay);
     }
@@ -99,13 +99,20 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
     @Override
     public RecordSuccessResponse recordSuccess(RecordSuccessRequest request) {
         DefaultRetryToken token = asStandardRetryToken(request.token());
+
+        // Update the circuit breaker token bucket.
         ReleaseResponse releaseResponse = updateCircuitBreakerTokenBucket(token);
+
+        // Refresh the retry token and return
         DefaultRetryToken refreshedToken = refreshRetryTokenAfterSuccess(token, releaseResponse);
+
+        // Log success and return.
         logRecordSuccess(token, releaseResponse);
         return RecordSuccessResponse.create(refreshedToken);
     }
 
     @Override
+    @ToBuilderIgnoreField({"DEFAULT_EXCEPTION_TOKEN_COST", "DEFAULT_TOKEN_BUCKET_SIZE"})
     public Builder toBuilder() {
         return new Builder(this);
     }
@@ -115,18 +122,6 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
      */
     public static Builder builder() {
         return new Builder();
-    }
-
-    private Duration computeBackoff(RefreshRetryTokenRequest request, DefaultRetryToken token) {
-        Duration backoff;
-        if (treatAsThrottling.test(request.failure())) {
-            backoff = throttlingBackoffStrategy.computeDelay(token.attempt());
-        } else {
-            backoff = backoffStrategy.computeDelay(token.attempt());
-        }
-        // Take the max delay between the suggested delay and the backoff delay.
-        Duration suggested = request.suggestedDelay().orElse(Duration.ZERO);
-        return maxOf(suggested, backoff);
     }
 
     private Duration maxOf(Duration left, Duration right) {
@@ -162,7 +157,7 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
                      .addFailure(failure)
                      .build();
             String message = acquisitionFailedMessage(acquireResponse);
-            LOG.debug(() -> message, failure);
+            LOG.error(() -> message, failure);
             throw new TokenAcquisitionFailedException(message, refreshedToken, failure);
         }
     }
@@ -179,7 +174,7 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
                      .addFailure(failure)
                      .build();
             String message = maxAttemptsReachedMessage(refreshedToken);
-            LOG.debug(() -> message, failure);
+            LOG.error(() -> message, failure);
             throw new TokenAcquisitionFailedException(message, refreshedToken, failure);
         }
     }
@@ -199,7 +194,8 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
                      .build();
             throw new TokenAcquisitionFailedException(message, refreshedToken, failure);
         }
-        LOG.debug(() -> nonRetryableExceptionMessage(token), failure);
+        int attempt = token.attempt();
+        LOG.warn(() -> String.format("Request attempt %d encountered retryable failure.", attempt), failure);
     }
 
     private String nonRetryableExceptionMessage(DefaultRetryToken token) {
@@ -267,17 +263,12 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
                                      token.getClass().getName());
     }
 
-    private AcquireResponse requestAcquireCapacity(RefreshRetryTokenRequest request, DefaultRetryToken token) {
+    private AcquireResponse requestAcquireCapacity(DefaultRetryToken token) {
         TokenBucket tokenBucket = tokenBucketStore.tokenBucketForScope(token.scope());
-        int amountToAcquire = 0;
-        if (circuitBreakerEnabled) {
-            if (treatAsThrottling.test(request.failure())) {
-                amountToAcquire = throttlingExceptionCost;
-            } else {
-                amountToAcquire = exceptionCost;
-            }
+        if (!circuitBreakerEnabled) {
+            return tokenBucket.tryAcquire(0);
         }
-        return tokenBucket.tryAcquire(amountToAcquire);
+        return tokenBucket.tryAcquire(exceptionCost);
     }
 
     private DefaultRetryToken refreshToken(RefreshRetryTokenRequest request, AcquireResponse acquireResponse) {
@@ -291,30 +282,29 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
                     .build();
     }
 
-    public static class Builder implements LegacyRetryStrategy.Builder {
+    public static class Builder implements StandardRetryStrategy.Builder {
+        private static final int DEFAULT_TOKEN_BUCKET_SIZE = 500;
         private List<Predicate<Throwable>> predicates;
-        private Integer maxAttempts;
-        private Boolean circuitBreakerEnabled;
-        private Integer exceptionCost;
-        private Integer throttlingExceptionCost;
-        private Predicate<Throwable> treatAsThrottling;
+        private int maxAttempts;
+        private boolean circuitBreakerEnabled;
+        private int exceptionCost;
         private BackoffStrategy backoffStrategy;
-        private BackoffStrategy throttlingBackoffStrategy;
         private TokenBucketStore tokenBucketStore;
 
         Builder() {
             predicates = new ArrayList<>();
+            circuitBreakerEnabled = true;
+            tokenBucketStore = TokenBucketStore.builder()
+                                               .tokenBucketMaxCapacity(DEFAULT_TOKEN_BUCKET_SIZE)
+                                               .build();
         }
 
-        Builder(LegacyRetryStrategyImpl strategy) {
+        Builder(DefaultStandardRetryStrategy strategy) {
             this.predicates = new ArrayList<>(strategy.predicates);
             this.maxAttempts = strategy.maxAttempts;
             this.circuitBreakerEnabled = strategy.circuitBreakerEnabled;
             this.exceptionCost = strategy.exceptionCost;
-            this.throttlingExceptionCost = strategy.throttlingExceptionCost;
-            this.treatAsThrottling = strategy.treatAsThrottling;
             this.backoffStrategy = strategy.backoffStrategy;
-            this.throttlingBackoffStrategy = strategy.throttlingBackoffStrategy;
             this.tokenBucketStore = strategy.tokenBucketStore;
         }
 
@@ -332,7 +322,11 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
 
         @Override
         public Builder circuitBreakerEnabled(Boolean circuitBreakerEnabled) {
-            this.circuitBreakerEnabled = circuitBreakerEnabled;
+            if (circuitBreakerEnabled == null) {
+                this.circuitBreakerEnabled = true;
+            } else {
+                this.circuitBreakerEnabled = circuitBreakerEnabled;
+            }
             return this;
         }
 
@@ -342,25 +336,8 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
             return this;
         }
 
-        @Override
-        public Builder treatAsThrottling(Predicate<Throwable> treatAsThrottling) {
-            this.treatAsThrottling = treatAsThrottling;
-            return this;
-        }
-
-        @Override
-        public Builder throttlingBackoffStrategy(BackoffStrategy throttlingBackoffStrategy) {
-            this.throttlingBackoffStrategy = throttlingBackoffStrategy;
-            return this;
-        }
-
         public Builder tokenBucketExceptionCost(int exceptionCost) {
             this.exceptionCost = exceptionCost;
-            return this;
-        }
-
-        public Builder tokenBucketThrottlingExceptionCost(int throttlingExceptionCost) {
-            this.throttlingExceptionCost = throttlingExceptionCost;
             return this;
         }
 
@@ -370,8 +347,8 @@ public final class LegacyRetryStrategyImpl implements LegacyRetryStrategy {
         }
 
         @Override
-        public LegacyRetryStrategyImpl build() {
-            return new LegacyRetryStrategyImpl(this);
+        public DefaultStandardRetryStrategy build() {
+            return new DefaultStandardRetryStrategy(this);
         }
     }
 }
